input moved back into options provider (behaviour too specific to generalise into a component)
