+++ internal/tui/components/OptionsProvider/config.go
@@ -2,6 +2,7 @@ package optionsProvider
 
 import (
 	"context"
+	"github.com/charmbracelet/bubbles/key"
 
 	"github.com/charmbracelet/bubbles/list"
 	tea "github.com/charmbracelet/bubbletea"
@@ -27,4 +28,17 @@ type ListConfig[T, U any] struct {
 
 	GetItemsFunc func(context.Context) ([]T, error)
 	// Style    lipgloss.Style
+}
+
+type InputConfig struct {
+	Prompt      string
+	Placeholder string
+	CharLimit   int
+	Width       int
+	KeyMap      InputKeyMaps
+}
+
+type InputKeyMaps struct {
+	Accept key.Binding
+	Back   key.Binding
 }
